v1.1.5 (r0030)
This should've been v1.2, but it is a bit unstable so until bugfixes
v1.2 isn't coming out. Added a Module containing the ResetConfig()
function (which really shortens FirstTimeSetupForm1.vb's code) and a
MouseClick() function (which simulates Mouse Click which is required if
you want to stop resizing). The resize is mastered. Added Settings
Window which allows changing OS and Size settings (if required).